+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ v1.9.1
 *Release date: In development*
 
 - Fix string conversion in dataset utilities
+- Add upper/lower conversion to replace param method
 
 v1.9.0
 ------